User Input Added
Now they will read the options (1,2,3,4) and press the corresponding key to indicate what they want. This causes the options to dissapear and the paragraph explaining why its right will appear. In the code Ayush I have indicated where you shoudl try to implement the second half of the video.